Rename mapper to project_mapper
